add more core modules
