+++ core/src/constants/index.spec.ts
