+++ core/src/constants/index.ts
