+++ core/src/export.ts
@@ -1,16 +1,16 @@
 export { isAdBlockInUse } from './detect-ad-blocker';
 export { EventTimer } from './event-timer';
-export { adSizes } from '../../src/lib/ad-sizes';
-export * as constants from '../../src/lib/constants';
+export { adSizes } from './ad-sizes';
+export * as constants from './constants';
 export {
 	bypassCommercialMetricsSampling,
 	initCommercialMetrics,
 } from '../../src/lib/send-commercial-metrics';
-export { buildPageTargeting } from '../../src/lib/targeting/build-page-targeting';
-export { postMessage } from '../../src/lib/messenger/post-message';
-export { buildImaAdTagUrl } from '../../src/lib/targeting/youtube-ima';
-export { getPermutivePFPSegments } from '../../src/lib/permutive';
-export type { AdsConfigDisabled } from '../../src/lib/types';
-export type { AdSize, SizeMapping, SlotName } from '../../src/lib/ad-sizes';
-export type { PageTargeting } from '../../src/lib/targeting/build-page-targeting';
-export type { AdsConfigUSNATorAus, AdsConfigTCFV2 } from '../../src/lib/types';
+// export { buildPageTargeting } from '../../src/lib/targeting/build-page-targeting';
+// export { postMessage } from '../../src/lib/messenger/post-message';
+// export { buildImaAdTagUrl } from '../../src/lib/targeting/youtube-ima';
+// export { getPermutivePFPSegments } from '../../src/lib/permutive';
+// export type { AdsConfigDisabled } from '../../src/lib/types';
+export type { AdSize, SizeMapping, SlotName } from './ad-sizes';
+// export type { PageTargeting } from '../../src/lib/targeting/build-page-targeting';
+// export type { AdsConfigUSNATorAus, AdsConfigTCFV2 } from '../../src/lib/types';
