+++ core/tsconfig.compile.json
@@ -1,0 +1,4 @@
+{
+	"extends": "./tsconfig.json",
+	"include": ["./src/export.ts", "./src/types.ts"]
+}
