+++ src/define/define-slot.spec.ts
@@ -1,5 +1,5 @@
-import type { SizeMapping } from '../lib/ad-sizes';
-import { adSizes, createAdSize } from '../lib/ad-sizes';
+import type { SizeMapping } from '@guardian/commercial/ad-sizes';
+import { adSizes, createAdSize } from '@guardian/commercial/ad-sizes';
 import {
 	buildGoogletagSizeMapping,
 	collectSizes,
