+++ src/display/request-bids.ts
@@ -1,5 +1,5 @@
 import type { Advert } from '../define/Advert';
-import type { AdSize } from '../lib/ad-sizes';
+import type { AdSize } from '@guardian/commercial/ad-sizes';
 import { a9 } from '../lib/header-bidding/a9/a9';
 import { prebid } from '../lib/header-bidding/prebid/prebid';
 import type { HeaderBiddingSlot } from '../lib/header-bidding/prebid-types';
