+++ src/events/on-slot-viewable.ts
@@ -3,8 +3,8 @@ import { log } from '@guardian/libs';
 import type { Advert } from '../define/Advert';
 import { isAdSize } from '../define/Advert';
 import { enableLazyLoad } from '../display/lazy-load';
-import { outstreamSizes } from '../lib/ad-sizes';
-import { AD_LABEL_HEIGHT } from '../lib/constants/ad-label-height';
+import { outstreamSizes } from '@guardian/commercial/ad-sizes';
+import { AD_LABEL_HEIGHT } from '@guardian/commercial/constants/ad-label-height';
 import { getAdvertById } from '../lib/dfp/get-advert-by-id';
 import { memoizedFetchNonRefreshableLineItemIds } from '../lib/dfp/non-refreshable-line-items';
 import { shouldRefresh } from '../lib/dfp/should-refresh';
