+++ src/events/render-advert.ts
@@ -1,5 +1,5 @@
 import type { Advert } from '../define/Advert';
-import { adSizes } from '../lib/ad-sizes';
+import { adSizes } from '@guardian/commercial/ad-sizes';
 import { reportError } from '../lib/error/report-error';
 import fastdom from '../lib/fastdom-promise';
 import { logGumGumWinningBid } from '../lib/gumgum-winning-bid';
