+++ src/init/consented/prepare-googletag.spec.ts
@@ -1,7 +1,7 @@
 import type { ConsentState, USNATConsentState } from '@guardian/libs';
 import { getConsentFor, loadScript, onConsent } from '@guardian/libs';
 import { loadAdvert } from '../../display/load-advert';
-import type * as AdSizesType from '../../lib/ad-sizes';
+import type * as AdSizesType from '@guardian/commercial/ad-sizes';
 import { commercialFeatures } from '../../lib/commercial-features';
 import { getCurrentBreakpoint as getCurrentBreakpoint_ } from '../../lib/detect/detect-breakpoint';
 import { dfpEnv } from '../../lib/dfp/dfp-env';
