+++ src/insert/fill-dynamic-advert-slot.ts
@@ -3,7 +3,7 @@ import type { Advert } from '../define/Advert';
 import { createAdvert } from '../define/create-advert';
 import { enableLazyLoad } from '../display/lazy-load';
 import { loadAdvert } from '../display/load-advert';
-import type { SizeMapping } from '../lib/ad-sizes';
+import type { SizeMapping } from '@guardian/commercial/ad-sizes';
 import { dfpEnv } from '../lib/dfp/dfp-env';
 import { queueAdvert } from '../lib/dfp/queue-advert';
 import { reportError } from '../lib/error/report-error';
