+++ src/lib/dfp/should-refresh.spec.ts
@@ -1,5 +1,5 @@
 import { Advert } from '../../define/Advert';
-import { adSizes, createAdSize } from '../../lib/ad-sizes';
+import { adSizes, createAdSize } from '@guardian/commercial/ad-sizes';
 import { shouldRefresh } from './should-refresh';
 
 const outstreamSizes = [
