+++ src/lib/dfp/should-refresh.ts
@@ -1,6 +1,6 @@
 import type { Advert } from '../../define/Advert';
-import type { AdSizeString } from '../../lib/ad-sizes';
-import { outstreamSizes } from '../../lib/ad-sizes';
+import type { AdSizeString } from '@guardian/commercial/ad-sizes';
+import { outstreamSizes } from '@guardian/commercial/ad-sizes';
 
 /**
  * Determine whether an advert should refresh, taking into account
