+++ src/lib/header-bidding/prebid/appnexus.spec.ts
@@ -1,4 +1,4 @@
-import { createAdSize } from '../../../lib/ad-sizes';
+import { createAdSize } from '@guardian/commercial/ad-sizes';
 import {
 	isInAuOrNz as isInAuOrNz_,
 	isInRow as isInRow_,
