+++ src/lib/header-bidding/prebid/prebid.ts
@@ -5,9 +5,9 @@ import { flatten } from 'lodash-es';
 import type { PrebidPriceGranularity } from 'prebid.js/src/cpmBucketManager';
 import type { Advert } from '../../../define/Advert';
 import { getParticipations } from '../../../experiments/ab';
-import type { AdSize } from '../../../lib/ad-sizes';
-import { createAdSize } from '../../../lib/ad-sizes';
-import { PREBID_TIMEOUT } from '../../../lib/constants/prebid-timeout';
+import type { AdSize } from '@guardian/commercial/ad-sizes';
+import { createAdSize } from '@guardian/commercial/ad-sizes';
+import { PREBID_TIMEOUT } from '@guardian/commercial/constants/prebid-timeout';
 import { getPermutiveSegments } from '../../../lib/permutive';
 import type { PageTargeting } from '../../../lib/targeting/build-page-targeting';
 import { pubmatic } from '../../__vendor/pubmatic';
