+++ src/lib/header-bidding/utils.ts
@@ -1,6 +1,6 @@
 import { type ConsentState, getConsentFor, isString } from '@guardian/libs';
 import { once } from 'lodash-es';
-import { createAdSize } from '../../lib/ad-sizes';
+import { createAdSize } from '@guardian/commercial/ad-sizes';
 import {
 	isInAuOrNz,
 	isInCanada,
